+++ backend/app/views/room.py
@@ -2,7 +2,7 @@ import math
 from flask import Blueprint, request, jsonify, session
 from sqlalchemy import desc
 from app import db
-from app.models import Status, Device
+from app.datamodel import Status, Device
 from app.utils import check_csrf_token
 
 room_blueprint = Blueprint("room", __name__)
